requirements.txt
Django==5.0.3
djangorestframework==3.14.0
django-filter==24.1
spacy==3.7.4
